Add and integrate gesture manager class
